Allow WorkflowImplementationOptions to be passed in TestWorkflowExtension
use setFailWorkflowExceptionTypes instead of timeout-setting
