+++ temporal-testing/src/test/java/io/temporal/testing/junit5/testWorkflowImplementationOptions/TestWorkflowImplementationOptionsViceVersa.java
@@ -18,46 +18,47 @@
  * limitations under the License.
  */
 
-package io.temporal.testing.junit5.workflowImplementationOptions;
+package io.temporal.testing.junit5.testWorkflowImplementationOptions;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import io.temporal.client.WorkflowOptions;
+import io.temporal.client.WorkflowStub;
+import io.temporal.testing.TestWorkflowEnvironment;
 import io.temporal.testing.TestWorkflowExtension;
 import io.temporal.testing.WorkflowInitialTime;
-import io.temporal.testing.internal.SDKTestOptions;
-import io.temporal.testing.junit5.workflowImplementationOptions.TestWorkflowImplementationOptionsCommon.HelloActivityImpl;
-import io.temporal.testing.junit5.workflowImplementationOptions.TestWorkflowImplementationOptionsCommon.HelloWorkflow;
-import io.temporal.testing.junit5.workflowImplementationOptions.TestWorkflowImplementationOptionsCommon.HelloWorkflowImpl;
+import io.temporal.testing.junit5.testWorkflowImplementationOptions.TestWorkflowImplementationOptionsCommon.HelloWorkflowImpl;
+import io.temporal.worker.Worker;
 import java.time.Instant;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
 @Timeout(value = 30, unit = TimeUnit.SECONDS)
-public class TestWorkflowImplementationOptions {
+public class TestWorkflowImplementationOptionsViceVersa {
 
   @RegisterExtension
   public static final TestWorkflowExtension testWorkflow =
       TestWorkflowExtension.newBuilder()
-          .registerWorkflowImplementationTypes(
-              SDKTestOptions.newWorkflowImplementationOptionsWithDefaultStartToCloseTimeout(),
-              HelloWorkflowImpl.class)
-          .setActivityImplementations(new HelloActivityImpl())
+          .registerWorkflowImplementationTypes(HelloWorkflowImpl.class)
           .setInitialTime(Instant.parse("2021-10-10T10:01:00Z"))
           .build();
 
   @Test
   @WorkflowInitialTime("2020-01-01T01:00:00Z")
-  public void extensionShouldLaunchTestEnvironmentWithWorkflowImplementationOptions(
-      HelloWorkflow workflow) {
+  public void extensionShouldNotBeAbleToCallActivityBasedOnMissingTimeouts(
+      TestWorkflowEnvironment testEnv, WorkflowOptions workflowOptions, Worker worker)
+      throws InterruptedException, ExecutionException, TimeoutException {
 
-    assertEquals(
-        String.format(
-            "Hello World from activity BuildGreeting and workflow HelloWorkflow with %s startToCloseTimeout",
-            SDKTestOptions.newWorkflowImplementationOptionsWithDefaultStartToCloseTimeout()
-                .getDefaultActivityOptions()
-                .getStartToCloseTimeout()),
-        workflow.sayHello("World"));
+    WorkflowStub cut =
+        testEnv.getWorkflowClient().newUntypedWorkflowStub("HelloWorkflow", workflowOptions);
+    cut.start("World");
+
+    CompletableFuture<String> resultAsync = cut.getResultAsync(String.class);
+    assertThrows(TimeoutException.class, () -> resultAsync.get(5, TimeUnit.SECONDS));
   }
 }
